20120808_3_1_0_the_silent_war
git-svn-id: http://bean-lai-dreamfactory-hallo.googlecode.com/svn/trunk@3433 6721f6c1-56d4-3193-6ef9-340426ecb113

+++ code/web/bops/src/main/java/com/doucome/corner/web/bops/action/zhe/TaokeReportSettleAjaxAction.java
@@ -13,7 +13,7 @@ import com.doucome.corner.web.bops.model.JsonModel;
 
 /**
  * 
- * @author shenjia.caosj 2012-4-22
+ * @author langben 2012-4-22
  *
  */
 @SuppressWarnings("serial")
@@ -51,10 +51,10 @@ public class TaokeReportSettleAjaxAction extends BopsBasicAction{
 //			json.setDetail("input settleTo is not right .") ;
 //			return SUCCESS ;
 //		}
-		List<Integer> idList = new ArrayList<Integer>() ;
+		List<Long> idList = new ArrayList<Long>() ;
 		for(String id : idArray){
 			if(StringUtils.isNumeric(id)){
-				idList.add(Integer.valueOf(id)) ;
+				idList.add(Long.valueOf(id)) ;
 			}
 		}
 		int count = ddzTaokeReportSettleService.updateTaokeSettleStatus(idList, status) ;
